# api/gen/swagger/api/tool.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/tool.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ToolService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/ping": {
      "get": {
        "operationId": "ToolService_Ping",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiEmpty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "ToolService"
        ]
      }
    },
    "/v1/tool/hash": {
      "post": {
        "operationId": "ToolService_Hash",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/apiHashRes"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apiHashReq"
            }
          }
        ],
        "tags": [
          "ToolService"
        ]
      }
    }
  },
  "definitions": {
    "apiEmpty": {
      "type": "object"
    },
    "apiHashReq": {
      "type": "object",
      "properties": {
        "hash": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "base64": {
          "type": "string"
        }
      }
    },
    "apiHashRes": {
      "type": "object",
      "properties": {
        "num": {
          "type": "string"
        },
        "hex": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}